fix: 구단소개, 구단연혁 컴포넌트 관심사 분리
